Applying the filters for GRA1, STA3, and LAP2.
It seems as if Python already has functions that apply the filters that you asked me to apply.  I was able to complete this fairly quickly.  However, I am not sure if the filters I used are the actual ones that you wanted me to, but they did filter the images.